feat(symthaea): 147 new tests across 7 suites + SIMD cosine + amortization tuning
Tasks A-G: end-to-end cognitive validation (12), Broca quality validation (17),
inference/REPL integration (45), external benchmark harness (12), voice pipeline
integration (14), humanoid integration (47). Added AVX2+FMA cosine_f32 with
scalar fallback. Increased amortization intervals for 9 consciousness subsystems
(spectral MIP 47→97, consciousness profile 23→47, epistemic gate 7→13, etc.).
New consciousness_engine.rs and ethics_engine.rs modules. Linter-applied fixes
across benchmarks, examples, and source files.

+++ symthaea/examples/phi_crossvalidation.rs
@@ -244,7 +244,7 @@ fn main() {
             let resonant_result = resonant_calc.compute(&topo.node_representations);
 
             real_phis.push(real_phi);
-            resonant_phis.push(resonant_result.phi as f64);
+            resonant_phis.push(resonant_result.phi);
         }
 
         let real_mean: f64 = real_phis.iter().sum::<f64>() / n_samples as f64;
